+++ README.md
@@ -62,8 +62,7 @@ S3_LOCAL_ENDPOINT=http://localhost:4569
 S3_ATTACHMENTS_BUCKET_NAME=local-uploads
 URL=http://localhost/3000
 BANNER_TABLE_NAME=local-banners
-REPORT_METADATA_TABLE_NAME=local-mcpar-reports-metadata
-REPORT_DATA_TABLE_NAME=local-mcpar-reports-data
+MCPAR_REPORT_TABLE_NAME=local-mcpar-reports
 TEMPLATE_BUCKET=local-uploads
 DISABLE_ESLINT_PLUGIN=true
 COGNITO_USER_POOL_ID=us-east-1_lerDvs4wn
