+++ services/app-api/package.json
@@ -20,6 +20,7 @@
     "@types/jest": "^27.4.0",
     "aws-lambda": "^1.0.7",
     "jest": "^27.4.7",
+    "ksuid": "^3.0.0",
     "serverless-associate-waf": "^1.2.1",
     "serverless-plugin-typescript": "^2.1.0",
     "ts-jest": "^27.1.3",
@@ -28,7 +29,6 @@
   "dependencies": {
     "aws-jwt-verify": "^3.1.0",
     "jwt-decode": "^3.1.2",
-    "uuid": "^7.0.3",
     "yup": "^0.32.11"
   },
   "jest": {
